Rename partition file; address review comments
